Remove duplicate shared secret derivation code
